Fixing discrepancy start coordinates
The coordinates from UniProtKB sometimes generates errors in the heat map bed file. These errors are detected in UCSC genome browser but not in IGV. The error is a discrepancy between the genomic start location of the protein and the genomic start location of the first exon. The discrepancy generates a non 0 blockStart coordinate for the first block in the bed file.
For instance:

chr2	55291714	55334327	A0A2R8Y7D4	1000	-	55291714	55334327	1.0,0.28308823529411764,0.28308823529411764	18	 63,130,106,265,71,200,83,214,92,145,186,143,277,161,164,141,127,233,	2,3883,4236,4544,9492,10158,11355,17095,17448,20720,24214,25492,25850,27129,30814,36580,40852,42380, 

The number 2 at the beginning of the third row represent the starting point of the first block and must be = 0 because the calculation is a subtraction between the genomic start coordinates of the protein (55291714 in this case) and the genomic start location of the first exon. 
The modification considers the genomic start location of the first exon as the starting point of the protein.